refactor(dat): extract Loadfile-Only mode to DatLoadfileOnlyComposer (#611 PR3)
Move ComposeLoadfileOnly and Loadfile-Only header building from
DatComposer to DatLoadfileOnlyComposer. DatComposer now delegates
non-profile LoadfileOnly mode to DatLoadfileOnlyComposer. The profile
path remains in DatComposer (depends on StandardRowValues, moves
with PR4). DatComposer.cs 501 to 431 lines. No behavior change.
